move feature families into mods
* slight factoring of query & user
