+++ xayn-ai/src/ltr/features/dataiku.rs
@@ -8,90 +8,7 @@
 
 use itertools::Itertools;
 use smallvec::{smallvec, SmallVec};
-use std::collections::{HashMap, HashSet};
-
-struct UserFeatures {
-    /// Entropy over ranks of clicked results.
-    click_entropy: f32,
-    /// Click counts of results ranked 1-2, 3-6, 6-10 resp.
-    click_counts: ClickCounts,
-    /// Total number of search queries over all sessions.
-    num_queries: usize,
-    /// Mean number of words per query.
-    words_per_query: f32,
-    /// Mean number of unique query words per session.
-    words_per_session: f32,
-}
-
-struct QueryFeatures {
-    /// Entropy over ranks of clicked results.
-    click_entropy: f32,
-    /// Number of terms.
-    num_terms: usize,
-    /// Average `n` where query is the `n`th of a session.
-    rank_per_session: f32,
-    /// Average number of occurrences per session.
-    occurs_per_session: f32,
-    /// Total number of occurrences.
-    num_occurs: usize,
-    /// Mean reciprocal rank of clicked results.
-    click_mrr: f32,
-    /// Average number of clicks.
-    avg_clicks: f32,
-    /// Average number of skips.
-    avg_skips: f32,
-}
-
-/// Calculate query features for the given query and historical search results of a user.
-fn query_features(history: &[SearchResult], query: Query) -> QueryFeatures {
-    let history_q = history
-        .iter()
-        .filter(|r| r.query_id == query.id)
-        .collect_vec();
-
-    let click_entropy = click_entropy(&history_q);
-
-    let occurs = history_q
-        .iter()
-        .map(|r| (r.session_id, r.query_counter))
-        .collect::<HashSet<_>>();
-
-    let num_occurs = occurs.len();
-
-    let rank_sum = occurs
-        .into_iter()
-        .map(|(_, query_count)| query_count as f32)
-        .sum::<f32>();
-    let rank_per_session = rank_sum / num_occurs as f32;
-
-    let num_sessions = history_q.iter().unique_by(|r| r.session_id).count() as f32;
-    let occurs_per_session = num_occurs as f32 / num_sessions;
-
-    let clicked = history_q
-        .iter()
-        .filter(|r| r.relevance > ClickSat::Low)
-        .collect_vec();
-    let click_mrr = mean_recip_rank(&clicked, None, None);
-
-    let avg_clicks = clicked.len() as f32 / num_occurs as f32;
-
-    let avg_skips = history_q
-        .into_iter()
-        .filter(|r| r.relevance == ClickSat::Skip)
-        .count() as f32
-        / num_occurs as f32;
-
-    QueryFeatures {
-        click_entropy,
-        num_terms: query.words.len(),
-        rank_per_session,
-        occurs_per_session,
-        num_occurs,
-        click_mrr,
-        avg_clicks,
-        avg_skips,
-    }
-}
+use std::collections::HashMap;
 
 /// Mean reciprocal rank of results filtered by outcome and a predicate.
 ///
@@ -105,7 +22,7 @@ fn query_features(history: &[SearchResult], query: Query) -> QueryFeatures {
 /// i.e. satisfying `pred` and matching `outcome`.
 ///
 /// The formula uses some form of additive smoothing with a prior 0.283 (see Dataiku paper).
-fn mean_recip_rank(
+pub(crate) fn mean_recip_rank(
     rs: &[impl AsRef<SearchResult>],
     outcome: Option<MrrOutcome>,
     pred: Option<FilterPred>,
@@ -134,32 +51,32 @@ fn mean_recip_rank(
     numer / denom
 }
 
-struct Query {
-    id: i32,
-    words: Vec<i32>,
+pub(crate) struct Query {
+    pub(crate) id: i32,
+    pub(crate) words: Vec<i32>,
 }
 
-struct SearchResult {
+pub struct SearchResult {
     /// Session identifier.
-    session_id: i32,
+    pub(crate) session_id: i32,
     /// User identifier.
-    user_id: i32,
+    pub(crate) user_id: i32,
     /// Query identifier.
-    query_id: i32,
+    pub(crate) query_id: i32,
     /// Day of the search, an integer between 1 and 27.
-    day: u8,
+    pub(crate) day: u8,
     /// Words of the query, each masked.
-    query_words: Vec<i32>,
+    pub(crate) query_words: Vec<i32>,
     /// URL of result, masked.
-    url: i32,
+    pub(crate) url: i32,
     /// Domain of result, masked.
-    domain: i32,
+    pub(crate) domain: i32,
     /// Relevance level of the result.
-    relevance: ClickSat,
+    pub(crate) relevance: ClickSat,
     /// Position among other results.
-    position: Rank,
+    pub(crate) position: Rank,
     /// Query count within session.
-    query_counter: u8,
+    pub(crate) query_counter: u8,
 }
 
 impl AsRef<SearchResult> for SearchResult {
@@ -172,7 +89,7 @@ impl AsRef<SearchResult> for SearchResult {
 ///
 /// Based on Yandex notion of dwell-time: time elapsed between a click and the next action.
 #[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Copy, Hash)]
-enum ClickSat {
+pub(crate) enum ClickSat {
     /// Snippet examined but URL not clicked.
     Skip,
     /// Snippet not examined.
@@ -186,7 +103,7 @@ enum ClickSat {
 }
 
 #[derive(Hash, PartialEq, Eq, PartialOrd, Ord, Clone, Copy)]
-enum Rank {
+pub(crate) enum Rank {
     First = 1,
     Second,
     Third,
@@ -233,7 +150,7 @@ fn seasonality(history: &[SearchResult], domain: i32) -> f32 {
 }
 
 /// Entropy over the rank of the given results that were clicked.
-fn click_entropy(results: &[impl AsRef<SearchResult>]) -> f32 {
+pub(crate) fn click_entropy(results: &[impl AsRef<SearchResult>]) -> f32 {
     let rank_freqs = results
         .iter()
         .filter_map(|r| {
@@ -252,87 +169,8 @@ fn click_entropy(results: &[impl AsRef<SearchResult>]) -> f32 {
         .sum()
 }
 
-/// Click counter.
-struct ClickCounts {
-    /// Click count of results ranked 1-2.
-    click12: u32,
-    /// Click count of results ranked 3-5.
-    click345: u32,
-    /// Click count of results ranked 6 upwards.
-    click6up: u32,
-}
-
-impl ClickCounts {
-    fn new() -> Self {
-        Self {
-            click12: 0,
-            click345: 0,
-            click6up: 0,
-        }
-    }
-
-    fn incr(mut self, rank: Rank) -> Self {
-        use Rank::*;
-        match rank {
-            First | Second => self.click12 += 1,
-            Third | Fourth | Fifth => self.click345 += 1,
-            _ => self.click6up += 1,
-        };
-        self
-    }
-}
-
-fn click_counts(results: &[SearchResult]) -> ClickCounts {
-    results
-        .iter()
-        .filter(|r| r.relevance > ClickSat::Low)
-        .fold(ClickCounts::new(), |counter, r| counter.incr(r.position))
-}
-
-/// Calculate user features for the given historical search results of a user.
-fn user_features(history: &[SearchResult]) -> UserFeatures {
-    // all queries over all sessions
-    let all_queries = history
-        .iter()
-        .map(|r| (r.session_id, r.query_id, &r.query_words, r.query_counter))
-        .collect::<HashSet<_>>();
-
-    let num_queries = all_queries.len();
-    let words_per_query = all_queries
-        .iter()
-        .map(|(_, _, words, _)| words.len())
-        .sum::<usize>() as f32
-        / num_queries as f32;
-
-    let words_by_session =
-        all_queries
-            .into_iter()
-            .fold(HashMap::new(), |mut words_by_session, (s, _, ws, _)| {
-                let words = words_by_session
-                    .entry(s)
-                    .or_insert_with(HashSet::<i32>::new);
-                words.extend(ws);
-                words_by_session
-            });
-
-    let num_sessions = words_by_session.len();
-    let words_per_session = words_by_session
-        .into_iter()
-        .map(|(_, words)| words.len())
-        .sum::<usize>() as f32
-        / num_sessions as f32;
-
-    UserFeatures {
-        click_entropy: click_entropy(history),
-        click_counts: click_counts(history),
-        num_queries,
-        words_per_query,
-        words_per_session,
-    }
-}
-
 #[derive(Clone, Copy, PartialEq, Eq, Hash)]
-enum MrrOutcome {
+pub(crate) enum MrrOutcome {
     Miss,
     Skip,
     Click,
@@ -340,7 +178,7 @@ enum MrrOutcome {
 
 /// Atomic features of which an aggregate feature is composed of.
 #[derive(Clone, Copy, PartialEq, Eq, Hash)]
-enum AtomFeat {
+pub(crate) enum AtomFeat {
     /// MRR for miss, skip, click.
     MeanRecipRank(MrrOutcome),
     /// MRR for all outcomes.
@@ -351,119 +189,7 @@ enum AtomFeat {
     SnippetQuality,
 }
 
-type FeatMap = HashMap<AtomFeat, f32>;
-
-/// Cumulated features for a given user.
-struct CumFeatures {
-    /// Cumulated feature for matching URL.
-    url: FeatMap,
-}
-
-/// Determines the cumulated features for a given search result.
-///
-/// These are given by sums of conditional probabilities:
-/// ```text
-/// sum{cond_prob(outcome, pred(r.url))}
-/// ```
-/// where the sum ranges over each search result `r` ranked above `res`. `pred` is the predicate
-/// corresponding to the cumulated feature, and `outcome` one of its specified atoms.
-fn cum_features(hist: &[SearchResult], res: SearchResult) -> CumFeatures {
-    let url = hist
-        .iter()
-        // if res is ranked n, get the n-1 results ranked above res
-        .filter(|r| {
-            r.session_id == res.session_id
-                && r.query_id == res.query_id
-                && r.query_counter == res.query_counter
-                && r.position < res.position
-        })
-        // calculate specified cond probs for each of the above
-        .flat_map(|r| {
-            let pred = FilterPred::new(UrlOrDom::Url(r.url));
-            pred.cum_atoms()
-                .into_iter()
-                .map(move |outcome| (outcome, cond_prob(hist, outcome, pred)))
-        })
-        // sum cond probs for each outcome
-        .fold(HashMap::new(), |mut cp_map, (outcome, cp)| {
-            *cp_map.entry(AtomFeat::CondProb(outcome)).or_default() += cp;
-            cp_map
-        });
-
-    CumFeatures { url }
-}
-
-/// Aggregate features for a given user.
-struct AggregFeatures {
-    /// Aggregate feature for matching domain.
-    dom: FeatMap,
-    /// Aggregate feature for matching domain over anterior sessions.
-    dom_ant: FeatMap,
-    /// Aggregate feature for matching URL.
-    url: FeatMap,
-    /// Aggregate feature for matching URL over anterior sessions.
-    url_ant: FeatMap,
-    /// Aggregate feature for matching domain and query.
-    dom_query: FeatMap,
-    /// Aggregate feature for matching domain and query over anterior sessions.
-    dom_query_ant: FeatMap,
-    /// Aggregate feature for matching URL and query.
-    url_query: FeatMap,
-    /// Aggregate feature for matching URL and query over anterior sessions.
-    url_query_ant: FeatMap,
-    /// Aggregate feature for matching URL and query over current session.
-    url_query_curr: FeatMap,
-}
-
-/// Calculate aggregate features for the given search result and history of a user.
-fn aggreg_features(hist: &[SearchResult], r: SearchResult) -> AggregFeatures {
-    let anterior = SessionCond::Anterior(r.session_id);
-    let current = SessionCond::Current(r.session_id);
-    let r_url = UrlOrDom::Url(r.url);
-    let r_dom = UrlOrDom::Dom(r.domain);
-
-    let pred_dom = FilterPred::new(r_dom);
-    let dom = aggreg_feat(hist, &r, pred_dom);
-    let dom_ant = aggreg_feat(hist, &r, pred_dom.with_session(anterior));
-
-    let pred_url = FilterPred::new(r_url);
-    let url = aggreg_feat(hist, &r, pred_url);
-    let url_ant = aggreg_feat(hist, &r, pred_url.with_session(anterior));
-
-    let pred_dom_query = pred_dom.with_query(r.query_id);
-    let dom_query = aggreg_feat(hist, &r, pred_dom_query);
-    let dom_query_ant = aggreg_feat(hist, &r, pred_dom_query.with_session(anterior));
-
-    let pred_url_query = pred_url.with_query(r.query_id);
-    let url_query = aggreg_feat(hist, &r, pred_url_query);
-    let url_query_ant = aggreg_feat(hist, &r, pred_url_query.with_session(anterior));
-    let url_query_curr = aggreg_feat(hist, &r, pred_url_query.with_session(current));
-
-    AggregFeatures {
-        dom,
-        dom_ant,
-        url,
-        url_ant,
-        dom_query,
-        dom_query_ant,
-        url_query,
-        url_query_ant,
-        url_query_curr,
-    }
-}
-
-fn aggreg_feat(hist: &[SearchResult], r: &SearchResult, pred: FilterPred) -> FeatMap {
-    let eval_atom = |atom_feat| match atom_feat {
-        AtomFeat::MeanRecipRank(outcome) => mean_recip_rank(hist, Some(outcome), Some(pred)),
-        AtomFeat::MeanRecipRankAll => mean_recip_rank(hist, None, Some(pred)),
-        AtomFeat::SnippetQuality => snippet_quality(hist, r, pred),
-        AtomFeat::CondProb(outcome) => cond_prob(hist, outcome, pred),
-    };
-    pred.agg_atoms()
-        .into_iter()
-        .map(|atom_feat| (atom_feat, eval_atom(atom_feat)))
-        .collect()
-}
+pub(crate) type FeatMap = HashMap<AtomFeat, f32>;
 
 /// Quality of the snippet associated with a search result.
 ///
@@ -474,7 +200,7 @@ fn aggreg_feat(hist: &[SearchResult], r: &SearchResult, pred: FilterPred) -> Fea
 /// |hist({Miss, Skip}, pred)|
 /// ```
 /// where the sum ranges over all result sets containing a URL matching `res.url`.
-fn snippet_quality(hist: &[SearchResult], res: &SearchResult, pred: FilterPred) -> f32 {
+pub(crate) fn snippet_quality(hist: &[SearchResult], res: &SearchResult, pred: FilterPred) -> f32 {
     let pred_filtered = hist.iter().filter(|r| pred.apply(r));
     let denom = pred_filtered
         .clone()
@@ -563,7 +289,7 @@ impl<'a> ResultSet<'a> {
 /// ```
 /// The formula uses some form of additive smoothing with `prior(Miss)` = `1` and `0` otherwise.
 /// See Dataiku paper. Note then the `sum` term amounts to `1`.
-fn cond_prob(hist: &[SearchResult], outcome: ClickSat, pred: FilterPred) -> f32 {
+pub(crate) fn cond_prob(hist: &[SearchResult], outcome: ClickSat, pred: FilterPred) -> f32 {
     let prior = if outcome == ClickSat::Miss { 1 } else { 0 };
 
     let filtered_by_pred = hist.iter().filter(|r| pred.apply(r)).collect_vec();
@@ -592,7 +318,7 @@ impl DocAddr {
 }
 
 #[derive(Clone, Copy)]
-enum UrlOrDom {
+pub(crate) enum UrlOrDom {
     /// A specific URL.
     Url(i32),
     /// Any URL belonging to the given domain.
@@ -601,7 +327,7 @@ enum UrlOrDom {
 
 /// Query submission timescale.
 #[derive(Clone, Copy)]
-enum SessionCond {
+pub(crate) enum SessionCond {
     /// Before current session.
     Anterior(i32),
     /// Current session.
@@ -612,14 +338,14 @@ enum SessionCond {
 
 /// Filter predicate representing a boolean condition on a search result.
 #[derive(Clone, Copy)]
-struct FilterPred {
+pub(crate) struct FilterPred {
     doc: UrlOrDom,
     query: Option<i32>,
     session: SessionCond,
 }
 
 impl FilterPred {
-    fn new(doc: UrlOrDom) -> Self {
+    pub(crate) fn new(doc: UrlOrDom) -> Self {
         Self {
             doc,
             query: None,
@@ -627,12 +353,12 @@ impl FilterPred {
         }
     }
 
-    fn with_query(mut self, query_id: i32) -> Self {
+    pub(crate) fn with_query(mut self, query_id: i32) -> Self {
         self.query = Some(query_id);
         self
     }
 
-    fn with_session(mut self, session: SessionCond) -> Self {
+    pub(crate) fn with_session(mut self, session: SessionCond) -> Self {
         self.session = session;
         self
     }
@@ -640,7 +366,7 @@ impl FilterPred {
     /// Lookup the atoms making up the aggregate feature for this filter predicate.
     ///
     /// Mapping of (predicate => atomic features) based on Dataiku's winning model (see paper).
-    fn agg_atoms(&self) -> SmallVec<[AtomFeat; 6]> {
+    pub(crate) fn agg_atoms(&self) -> SmallVec<[AtomFeat; 6]> {
         use AtomFeat::{
             CondProb,
             MeanRecipRank as MRR,
@@ -673,7 +399,7 @@ impl FilterPred {
     ///
     /// Mapping of (predicate => atomic features) based on Dataiku's winning model (see paper).
     /// Note that for cumulated features, the atoms are always conditional probabilities.
-    fn cum_atoms(&self) -> SmallVec<[ClickSat; 3]> {
+    pub(crate) fn cum_atoms(&self) -> SmallVec<[ClickSat; 3]> {
         use ClickSat::{High as click2, Medium as click1, Skip as skip};
         use SessionCond::All;
         use UrlOrDom::*;
@@ -685,7 +411,7 @@ impl FilterPred {
     }
 
     /// Applies the predicate to the given search result.
-    fn apply(&self, r: impl AsRef<SearchResult>) -> bool {
+    pub(crate) fn apply(&self, r: impl AsRef<SearchResult>) -> bool {
         let r = r.as_ref();
         let doc_cond = match self.doc {
             UrlOrDom::Url(url) => r.url == url,
