+++ xayn-ai/src/ltr/features/mod.rs
@@ -1,1 +1,5 @@
+mod aggregate;
+mod cumulate;
 mod dataiku;
+mod query;
+mod user;
